phase2/ingestion_service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
google-cloud-storage>=2.10.0
